biocomputing/bioarm/pyproject.toml
[project]
name = "bioarm"
version = "0.1.0"
description = "BioARM project"
requires-python = ">=3.10"
dependencies = [
    "grpcio-tools==1.60.0",
    "grpcio==1.60.0",
    "protobuf==4.25.1",
    "libroadrunner==2.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true
